+++ src/main/java/org/yearup/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import org.yearup.models.Product;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer>
+public interface ProductRepository extends JpaRepository<Product, Long>
 {
-    List<Product> findByCategoryId(int categoryId);
+    List<Product> findByCategoryId(Long categoryId);
 }
